[FrameworkBundle] Add file links to named controllers in debug:router
