+++ src/Symfony/Bundle/FrameworkBundle/Command/ContainerDebugCommand.php
@@ -12,8 +12,6 @@
 namespace Symfony\Bundle\FrameworkBundle\Command;
 
 use Symfony\Bundle\FrameworkBundle\Console\Helper\DescriptorHelper;
-use Symfony\Component\Config\ConfigCache;
-use Symfony\Component\Config\FileLocator;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Exception\InvalidArgumentException;
 use Symfony\Component\Console\Input\InputArgument;
@@ -21,10 +19,8 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
-use Symfony\Component\DependencyInjection\Compiler\ServiceLocatorTagPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
-use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBag;
 
 /**
@@ -36,12 +32,9 @@ use Symfony\Component\DependencyInjection\ParameterBag\ParameterBag;
  */
 class ContainerDebugCommand extends Command
 {
-    protected static $defaultName = 'debug:container';
+    use BuildDebugContainerTrait;
 
-    /**
-     * @var ContainerBuilder|null
-     */
-    protected $containerBuilder;
+    protected static $defaultName = 'debug:container';
 
     /**
      * {@inheritdoc}
@@ -217,34 +210,6 @@ EOF
         } elseif ((null === $name) && $optionsCount > 1) {
             throw new InvalidArgumentException('The options tags, tag, parameters & parameter can not be combined together.');
         }
-    }
-
-    /**
-     * Loads the ContainerBuilder from the cache.
-     *
-     * @throws \LogicException
-     */
-    protected function getContainerBuilder(): ContainerBuilder
-    {
-        if ($this->containerBuilder) {
-            return $this->containerBuilder;
-        }
-
-        $kernel = $this->getApplication()->getKernel();
-
-        if (!$kernel->isDebug() || !(new ConfigCache($kernel->getContainer()->getParameter('debug.container.dump'), true))->isFresh()) {
-            $buildContainer = \Closure::bind(function () { return $this->buildContainer(); }, $kernel, \get_class($kernel));
-            $container = $buildContainer();
-            $container->getCompilerPassConfig()->setRemovingPasses([]);
-            $container->getCompilerPassConfig()->setAfterRemovingPasses([]);
-            $container->compile();
-        } else {
-            (new XmlFileLoader($container = new ContainerBuilder(), new FileLocator()))->load($kernel->getContainer()->getParameter('debug.container.dump'));
-            $locatorPass = new ServiceLocatorTagPass();
-            $locatorPass->process($container);
-        }
-
-        return $this->containerBuilder = $container;
     }
 
     private function findProperServiceName(InputInterface $input, SymfonyStyle $io, ContainerBuilder $builder, string $name, bool $showHidden): string
